requirements.txt
customtkinter
Pillow
requests
pytest
